Move unmount volume util from pkg/volume/util to pkg/util/mount
